Cleanup: remove dead exports, fix wasm init, move PropResult
- Remove dead ExprResult/isUnsupported from pant-ast.ts
- Fix bodyExpr to use proper type narrowing instead of unsafe cast
- Replace fragile setTimeout(200) wasm init with exponential-backoff
  polling (1ms-100ms, 10s timeout); remove redundant wasmLoaded flag
- Move PropResult from pant-ast.ts to types.ts (translation concept,
  not AST concept)
- Remove redundant section comments from pant_wasm.ml
